+++ src/analysis.c
@@ -133,7 +133,7 @@ double getBeatError(const struct myarr* totaltick, double rate, int verbose)
         syncwrite(totaltick->arr, ArrayLength / 2, "t1");
         syncwrite(totaltick->arr + ArrayLength / 2, ArrayLength / 2, "t2");
     }
-    unsigned int postick = getmaxpos(cross, ArrayLength / 2);
+    size_t postick = getmaxpos(cross, ArrayLength / 2);
     free(cross);
     return shiftHalf(postick, ArrayLength / 2) * THOUSAND / rate;
 }
